Add files via upload
This file is the result of my team's participation in VanSASH, a one day Datathon collaborative event between SFU's Sports Analytics Club and the Vancouver Whitecaps of MLS. We were unable to get permission to use the data provided to us during the hackathon, which includes any results as well. However, I am still uploading it with the hopes of someday finding a sufficient data set to apply our project to.